+++ src/phone.hpp
@@ -121,7 +121,7 @@ public:
     void MouseMoved(sf::Vector2i Pos);
     void LeftMouseClicked(sf::Vector2i Pos);
     void RightMouseClicked(SGInterpreter* pInterpreter);
-    void SetPhoneCallAllowMask(uint8_t Mask);
+    void SetPhoneCallAllowMask(uint16_t Mask);
     void AddressMenuHighlight(int32_t Index);
     void PhoneSendMailEdit(const string& Subject, const string& Sender, const string& Body);
